Move `logServeMode()` to `app-utils.js`
